REST - move data classes to models.kt
